modification et changement de status d'un compte administrateur

+++ src/app/components/administrateur/liste-user/liste-user.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { Administrateur } from 'src/app/models/administrateur.model';
 import { AdminService } from 'src/app/services/administrateurService/admin.service';
+import { EtatCompte } from '../../enums/etatCompte.enumm';
 
 @Component({
   selector: 'app-liste-user',
@@ -34,6 +36,8 @@ export class ListeUserComponent implements OnInit {
 
   /*La methode pour supprimer un admin */
   public delete(id:number){
+    let confirmation=confirm("La suppression d'un administrateur est irreversible etes-vous sur ?")
+    if(confirmation==false)return;
     this.serviceAdmin.deleteAdminstrateur(id)
       .subscribe({
         next:(data)=>{
@@ -44,6 +48,28 @@ export class ListeUserComponent implements OnInit {
           this.messageError=error
         }
       })
+  }
+
+  activeDesactive(admin:Administrateur){
+    if(admin.etatCompte==="ACTIVE"){
+      this.serviceAdmin.desctiveCompte(admin).subscribe({
+        next:(data)=>{
+          this.getAll();
+        },
+        error:(error)=>{
+          console.log(error); 
+        }
+      })
+    }else{
+      this.serviceAdmin.activeCompte(admin).subscribe({
+        next:(data)=>{  
+          this.getAll();
+        },
+        error:(error)=>{
+          console.log(error); 
+        }
+      })
+    }
   }
 
 }
